loky bot/pyproject.toml
[project]
name = "loky-bybit-futures-bot"
version = "3.0.0"
description = "Multi-strategy directional trading bot for Bybit USDT-M Linear Perpetual Futures"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "orjson>=3.11.7",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "websockets>=16.0",
    "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
